move hamiltonian database to dataset module

+++ nablaDFT/dataset/nablaDFT_dataset.py
@@ -14,10 +14,7 @@ from schnetpack.data import AtomsDataFormat, load_dataset
 import nablaDFT
 from nablaDFT.dataset.atoms_datamodule import AtomsDataModule
 
-
-
-from nablaDFT.phisnet.training.hamiltonian_dataset import HamiltonianDataset
-from nablaDFT.phisnet.training.sqlite_database import HamiltonianDatabase
+from nablaDFT.dataset import HamiltonianDatabase, HamiltonianDataset
 
 
 class ASENablaDFT(AtomsDataModule):
@@ -108,12 +105,16 @@ class HamiltonianNablaDFT(HamiltonianDataset):
         self.dtype = dtype
         if not os.path.exists(datapath):
             os.makedirs(datapath)
-        f = open(nablaDFT.__path__[0] + "/links/hamiltonian_databases.json")
-        data = json.load(f)
-        url = data["train_databases"][dataset_name]
-        f.close()
+        with open(nablaDFT.__path__[0] + "/links/hamiltonian_databases.json") as f:
+                data = json.load(f)
+                url = data["train_databases"][dataset_name]
+        # f = open(nablaDFT.__path__[0] + "/links/hamiltonian_databases.json")
+        # data = json.load(f)
+        # url = data["train_databases"][dataset_name]
+        # f.close()
         filepath = datapath + "/" + dataset_name + ".db"
-        request.urlretrieve(url, filepath)
+        if not os.path.exists(filepath):
+            request.urlretrieve(url, filepath)
         self.database = HamiltonianDatabase(filepath)
         max_orbitals = []
         for z in self.database.Z:
